refactor: Aside 내부의 CategoryNode 컴포넌트 분리
- Aside.styles.ts에서 CategoryNode 관련 스타일 분리
- Aside 컴포넌트에서 CategoryNode 관련 파일을 CategoryNode.style.ts로 분리
  - CategoryNode.tsx, CategoryNode.style.ts를 Aside/components 디렉토리로 이동
- 변경된 파일 구조에 맞게 import 경로 수정
