+++ FRONT/src/components/Aside/Aside.tsx
@@ -1,4 +1,4 @@
-import CategoryNode from './CategoryNode';
+import CategoryNode from './components/CategoryNode';
 import cancelButton from '../../assets/icons/cancel-button.svg'
 import logoIcon from '../../assets/logo/logo.svg';
 import useAside from './useAside';
